backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0

# Email
sendgrid==6.11.0

# Scheduler para jobs de limpeza
apscheduler==3.10.4

# HTTP Client
httpx==0.27.0

# Excel
openpyxl==3.1.2
pandas==2.1.3

# Templates
Jinja2==3.1.2

# HTML Parsing
beautifulsoup4==4.12.3

# Testing (apenas desenvolvimento)
# pytest==7.4.3
# pytest-asyncio==0.21.1

